refactor: moves wrapper classes and interfaces to their own files
